+++ Server2/Method.cs
@@ -1,0 +1,12 @@
+﻿namespace Server
+{
+    static class Method
+    {
+        public const string Read = "read";
+        public const string Create = "create";
+        public const string Update = "update";
+        public const string Delete = "delete";
+
+        public static readonly string[] methods = { Read, Create, Update, Delete };
+    }
+}
