+++ Server2/Response.cs
@@ -1,0 +1,8 @@
+﻿namespace Server
+{
+    public class Response
+    {
+        public string Status { get; set; }
+        public string Body { get; set; }
+    }
+}
